requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.2
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
packaging==24.1
pip-review==1.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
service-identity==24.1.0
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zope.interface==6.4.post2
celery~=5.4.0